+++ Engine/Item/Equipment.h
@@ -6,11 +6,11 @@
 class Equipment
 {
 public:
-    int hp = 0;
-    int min_dmg = 0;
-    int max_dmg = 0;
-    int crit = 0;
-    float attack_speed = 0.0; // 0% reduction
+    float hp = 0.f;
+    float min_dmg = 0.f;
+    float max_dmg = 0.f;
+    float crit = 0.f;
+    float attack_speed = 0.f; // 0% reduction
 };
 
 #endif // EQUIPMENT_H
